Split xslt transformer
Split the xslt transformers into dedicated transfromers for each document type using a parent class for common logic.
